feat(ir): P40-d — u8→u16 width migration + u128 Const payload
Widen every bit-vector *width* type from `u8` to `u16` and the `Const`
payload from `u64` to `u128`, so vector registers past xmm (ymm=256,
zmm=512) and >64-bit SIMD immediates become representable. Foundational
for P40-e (ymm/zmm layouts + AVX lifters). Atomic: the tree does not
compile until the change is applied consistently across all crates.

Width type (`Var.bits`, `Const.bits`, `Extract.hi/lo`, `*Extend.to_bits`,
`LoadMem/StoreMem.bits`, `RegisterLayout.{lo,hi,width()}`, pointer width,
and every `operand_width`/`binop_width`/`memory_operand_width` helper)
→ `u16`. Register *indices*, instruction byte sizes, tbz bit positions,
and the Thumb hint stay `u8`. Not a JSON wire break — serde emits both
widths as bare integers, and no width field is stringified.

`Const.value` → `u128`; the Z3 encoder gains `bv_const`, which assembles
a >64-bit literal from its two 64-bit halves via `Concat` (z3 has no
`from_u128`). The SMT-LIB backend renders the decimal `u128` directly.

Latent truncation bugs the migration exposes, fixed here (a pure rename
would be silently wrong):
- `simplify::width_mask` → `u128`, guarded at 128. The all-ones absorbing
  rule (`va == width_mask(bits)`) no longer misreads a 128-bit const whose
  low 64 bits are all-ones as all-ones-at-128 — that was a fabricated
  verdict. `Concat` width math widened (no longer `None` at 128+128).
- P-code `INT_NEGATE` builds all-ones at the real output width, not
  `u64::MAX`; `varnode_bits` widens byte→bit before the multiply.
- ESIL load/store byte→bit conversion widens before clamping.
- SMT-LIB store byte count no longer saturates at 255 bits.

Regression artifacts:
- `r2smt-ir`: zmm-width var + `Extract{511,256}` + a 128-bit all-ones
  `Const` round-trip through JSON unchanged.
- `r2smt-core`: a 128-bit all-ones `Const`'s bit 127 solves to 1
  (`AlwaysTrue`) — proves the high half survives encoding, which a
  `BV::from_u64`-only path would truncate to `AlwaysFalse`.

Gates green: fmt, clippy -D warnings, test --all, solver-contracts,
determinism. (deny/audit's `atty` advisory is pre-existing via radius2,
unrelated to this change.)

+++ crates/r2smt-common/src/types.rs
@@ -99,7 +99,7 @@ pub enum Arch {
 impl Arch {
     /// Default register width in bits.
     #[must_use]
-    pub const fn pointer_bits(self) -> u8 {
+    pub const fn pointer_bits(self) -> u16 {
         match self {
             Self::X86 | Self::Arm => 32,
             Self::X86_64 | Self::Aarch64 => 64,
